Move lualine configuration up to common, fix lualine session component, remove tabline.nvim and replace enable_tabline with enable_status_in_tab
