+++ lua/free/plugins/ui.lua
@@ -67,7 +67,7 @@ if settings.enable_statusline then
       },
       config = function(_, opts)
         if settings.lualine_style == "onno" then
-          local lualine_config = require("free.config.lualine")
+          local lualine_config = require("config.lualine")
           lualine_config.setup(opts)
           lualine_config.load()
         else
@@ -87,7 +87,7 @@ if settings.enable_statusline then
       },
       config = function(_, opts)
         if settings.lualine_style == "onno" then
-          local lualine_config = require("free.config.lualine")
+          local lualine_config = require("config.lualine")
           lualine_config.setup(opts)
           lualine_config.load()
         else
@@ -96,17 +96,6 @@ if settings.enable_statusline then
       end,
     }
   end
-end
-
-local tabline_cfg = {}
-if settings.enable_tabline then
-  tabline_cfg = {
-    "kdheepak/tabline.nvim",
-    event = "VeryLazy",
-    config = function()
-      require("config.tabline")
-    end,
-  }
 end
 
 local indentline_cfg = {
@@ -361,9 +350,6 @@ return {
 
   -- LSP progress in lualine
   lualine_lsp_progress,
-
-  -- tabline
-  tabline_cfg,
 
   -- indent guides for Neovim
   indentline_cfg,
